backend/requirements.txt
Flask==2.0.1
flask-cors==3.0.10
requests==2.26.0
python-dateutil==2.8.2
openai==0.27.0
